Fix startup hang
The startup hang is due to the CEF's initialization in
BrowserManager::Impl::BrowserManagerEntry() taking much longer than it
used to take, so long that a browser source may try to be created in
BrowserManager::Impl::CreateBrowser before CEF is actually finished
initializing.  When that happens and a browser source tried to be
created, the CefPostTask that's used in source creation would fail and
not properly post/execute the task, causing the browser source to
indefinitely hang while waiting for the task to finish, when in fact it
never executes at all.

To fix this issue, a manual reset startup event is used, which is
signaled when CEF finishes initialization.  Then when trying to create
any browser source, it will wait to make sure that CEF startup event has
been signaled before attempting to post a task in to the Cef message
queue.

This has been a latent bug that had only been exposed recently, most
likely due to recent CEF changes that caused it to take longer than
normal to initialize.

Additionally, this bug most likely commonly happened once after system
startup due to caching of some sort.