Fix BrowserClient accessing data after destroy
The BrowserClient is a reference-counted object, and is a separate
object from both the BrowserSource object and the
BrowserSource::cefBrowser object.  When
cefBrowser->GetHost()->CloseBrowser(true) is called in
BrowserSource::DestroyBrowser(), the BrowserClient object is not
necessarily guaranteed to have been destroyed, and may still have
activity queued in the CEF thread.  Thus it becomes important to check
to make sure that the source is still available when executing any
functions in the BrowserClient object.  A more ideal solution would be
to make the BrowserSource object a atomically reference counted pointer
at some point.

Closes obsproject/obs-browser#113