Change options to be easier to use with config
This removes the "_ENABLED" suffix on options. The advantage of this
is it makes more sense and it's easier to work with from CMake. The
script will now check to see if that EXPERIMENTAL_SHARED_TEXTURE_SUPPORT
is true. If it is, it sets the enabled variable which can then be used
in code correctly with the need to define macros that hope to catch the
value of that option. This way, any false or true value that cmake knows
will correctly work. If one is given to cmake that it doesn't understand
it will correctly tell the user.