Blacklist certain devices from hardware accel
When using more than one adapter and the Intel adapter is in use
(basically, when the user is using the wrong device on a laptop), do not
use hardware acceleration for the browser source.  When the main program
is running on the Intel adapter and the browser source is running on
the NVIDIA/AMD adapter, the browser source will not render with hardware
acceleration.