Add option to use custom framerate
Due to the shared texture work done for OBS 22 and a define being true
even when not running obs-browser with hardware acceleration, the FPS of
the browser source would always be tied to the framerate for the OBS
output. This differs from behavior of previous versions where the FPS
would always be what the user entered as their desired FPS.

As a result, the FPS input for the Browser Source is currently silently
ignored. This also can cause problems for browser graphics that have
been coded with a specific FPS intended for it to run at regardless of
OBS' own framerate.

This change will mean that we still get the benefit of allowing the
browser to be tied to OBS' framerate and cut down on excessive renders
in those situations, while allowing an override if the user wishes for
it.