Enable OBS Studio service integration support (#67)
* Fix getCurrentScene returning source information instead of scene info

* Fix setting cache_path in portable mode

cache_path seems to require an absolute path

* Fix crash when no current scene is set in the scene collection

* Change cef-bootstrap to a more descriptive name

* Include string.h to have strlen available on Linux

* Fix memory leak

Fixes a memory leak indroduced by 2efce989

* Blacklist a few more devices from browser hwaccel

Blacklists Intel devices in general, and blacklists Microsoft's basic
render driver.

* Fix render bug when using OpenGL and hwaccel

If hardware acceleration is not available, and shared texture
functionality is not available, it would still have 'hwaccel' as true,
which would affect other parts of the program.  Instead, just set
hwaccel to false if texture sharing is not available.

* Update version to 2.1.3

* Add option to use custom framerate

Due to the shared texture work done for OBS 22 and a define being true
even when not running obs-browser with hardware acceleration, the FPS of
the browser source would always be tied to the framerate for the OBS
output. This differs from behavior of previous versions where the FPS
would always be what the user entered as their desired FPS.

As a result, the FPS input for the Browser Source is currently silently
ignored. This also can cause problems for browser graphics that have
been coded with a specific FPS intended for it to run at regardless of
OBS' own framerate.

This change will mean that we still get the benefit of allowing the
browser to be tied to OBS' framerate and cut down on excessive renders
in those situations, while allowing an override if the user wishes for
it.

* Do not handle scene change event if source nameless

* Use OBSSource helper to ensure safety

* Check valid scene for getCurrentScene

* Don't use texture copy for hwaccel

This was an unnecessary thing to do, and can cause the texture to get
stuck in a mid-render step.

* Fix crash on startup on macOS

Getting the absolute path on macOS requires that the relative
path exists.

* Update version to 2.1.6

* Pass Through argc/argv On Linux/macOS

Use the new `obs_get_cmdline_args()` from libobs to actually initialize CEF with
argc/argv on Linux/macOS platforms -- this previously only worked on Windows.
I've tested this with `--enable-media-stream` and `--enable-gpu` with success.

* Use baked CEF dir on OSX unless deploying

Makes it so it's easier to compile/use the browser plugin on mac.

* Don't find debug CEF on non-windows systems

* Parse json before sending data to events

* Add high DPI awareness to client and subprocesses

Fixes a bug where the browser widget/panel would sometimes render at an
incorrect size on high DPI monitors.

* Move string variable inside of if statement

This variable wasn't being used outside of the if statement, so just
move it inside.

* Add QCefWidget::urlChanged signal

Allows the ability for the frontend to be able to know when the user has
browsed to another URL in the browser widget.

* Fix building with CEF versions higher than or equal to 3578

* Fixed compilation with Chromium < 3578

Regression from 835b5418bfe77e5338a0322365011f0761bc1048

* Fix issue when trying to load a non-existing local file
If a non-existing local file was encountered, the entire request
processing would be cancelled.

* Add event to determine when CEF can post tasks

* Add Radeon HD 8850M to HW accel blacklist

* Update README.md

* Add Radeon HD 7660 to HW accel blacklist

* Fix multi-monitor multi-DPI rendering issues

* Remove CEF file logging

* Add current OBS locale to accepted language list

* Add Ctrl-R shortcut to refresh panel

* Open tabbed (or middle-click) URLs in sys. browser

When a user middle-clicked a URL, it would open that URL in the current
panel.  This fixes that by using the OnOpenURLFromTab callback to open
the URL in the user's system browser instead of the main program.

* Add ability to set panel startup script

Allows the ability to execute scripts on page startup.

* Make sure browser's not created in create call

It's possible for the create task to be queued more than once before the
first call has been executed, so make sure to check that the cefBrowser
member object hasn't already been created when this function is
executed.

* Add new panel API

Uses a new export, which returns a pointer to a pure virtual abstract
class in order make it a little bit easier to add new functions.

* Add cookie manager

Allows the ability to manage cookies separately for specific panels, or
to swap out cookies depending on the situation (such as switching
profiles).

* Add ability to whitelist specific popups

Allows the ability to whitelist specific CEF popups for specific
QObjects.

* Enable browser panel support in build script