Fix browser not always initially showing
Fixes a bug where the browser would initially start up as blank/white.
This is likely a bug internally with chromium (at least this version of
chromium) or CEF on high-DPI displays.  Posting a WM_SIZE message tells
Chromium/CEF that the window position/size has changed, and it adjusts
it back in to the right position.