Fix windows browser deadlock
While blocking the main application UI thread to wait on a browser's
task in the browser UI thread, the browser may call SendMessage on the
main UI thread, causing it to block and wait for the UI thread,
resulting in a deadlock.  This solves that by occasionally calling
QCoreApplication::processEvents to make it process UI messages/events
while waiting for the task to complete.