Revert use of OBS FPS instead of SignalBeginFrame
Because of the issues with the WasHidden() call in CEF version 3507 and
greater, workarounds were attempted in order to solve it.  One of those
workarounds was instead of using SignalBeginFrame(), run the browsers at
OBS FPS at all times.

Unfortunately, this workaround may have solved issues with WasHidden()
in some cases, but it did not solve all of them.  Problems still
occurred with WasHidden() usage.  Therefore, the logical solution is to
remove WasHidden() instead.  Thus, the workaround code to use OBS FPS in
place of SignalBeginFrame() no longer has purpose.