Use container widget for CEF
During testing, it was discovered that calling DestroyWindow() on the
browser's HWND doesn't actually destroy the browser.  Internally, it
will detect that the window handle was forcibly closed, but because the
parent still exists, it will just recreate it instantly.  This behavior
was noted when attempting to shut down the browser before destroying the
widget itself.  The browser kept going and audio kept playing.

So to fix this, there needs to be yet another widget inside of the
widget to contain CEF, which must be manually destroyed when we want to
destroy the browser, and recreated when we want a new browser.  Doing
this solves the issue and seems to fully destroy the browser internally
within CEF.