Remove unnecessary call to SetDocumentVisibility
These functions were added to emulate visibility for browser sources,
but calling it in CefRenderProcessHandler::OnContextCreated() will cause
it to affect browser panels as well.  This made it so that their initial
state was hidden, causing any video feed (particularly twitch) to not
start up on browser panels.

This function call was unnecessary anyway due to the fact that when a
browser source is created, SendBrowserVisibility(cefBrowser, is_showing)
is called subsequently thereafter.