Fix initial browser source visibility state
Fixes the initial visibility state of browser sources.  Removing the
visibility call in OnContextCreated made it so that the initial
visibility state would always end up being true.  It was originally
removed because the initial visibility state was always set to true,
even for browser panels.  Initially, it was thought that the code was
redundant, but it turns out the frame can be initially created then
destroyed, then recreated again for whatever reason, and thus the
last visibility state has to be stored in order to ensure that it stays
visible/invisible.

This fixes that by setting the visibility only if visibility has been
explicitly set previously by a visibility message from the application.
It stores the visibility state in an unordered_map if set, and then if
OnContextCreated is called again at a later time with a new frame for
whatever reason, it will be able to see that the value was actually
explicitly set, thus this fixes the issue of it affecting browser
panels, which do not send that message.