Disable rerouting by default
It was decided that due to the fact that audio monitoring has not had
testing on this wide of a scale of users before, that holding off on
rerouting audio is the best interim solution -- make it optional, and
off by default.  Users can still use this, but they will need to
explicitly enable it themselves.