Use OBS canvas size as browser source default size
This is a minor QoL change to make browser sources default to be the
same size as the user's OBS canvas. For users using overlays, this
should reduce the work required to add a browser source for an overlay.
For all other use cases, this should not create extra work.