Use older chromium network implementation
This uses the older network implementation rather than the new
NetworkService chromium has implemented until a later version of CEF is
used.  The newer NetworkService implementation has caused some lag with
playback when hardware acceleration is not available for whatever
reason.  This doesn't fix media playback issues with software rendering
on windows, but it does on macOS.  All playback issues with software
rendering appear to be fixed in later version of CEF, however those
versions do not currently have hardware accelerated OSR or audio capture
support, so we must deal with 3770 until we can upgrade CEF again (which
hopefully won't be too long).