Remove "monitor by default" flag
The reasoning behind removing this flag is because the whole point of
having it in was so we could replace chromium's audio output.  But there
are too many obstacles in our path from doing that for now, so it's
better to not have this flag because it just causes a lot of browser
sources to initialize audio monitoring unnecessarily.  We can change
this in the future, but for now, it's best to let the user choose to
turn on audio monitoring for the source if they choose to turn on audio
rerouting to OBS.

Note that this only changes default behavior for newly created sources;
it will not change settings of existing sources the user had before this
change.