Remove ExecuteOnBrowser wrapper on destroy
Under certain conditions, when browser source scene items are removed
in a loop, CEF UI thread may hang waiting for previous invocations of
ExecuteOnBrowser() to complete resulting in a deadlock.

Since CefBrowserHost::WasHidden() and CefBrowserHost::CloseBrowser()
are not required to be called on a certain thread, removing
ExecuteOnBrowser() altogether preserves the functionality, eliminates
the deadlock and simplifies the code.