Fix Chromium modifying process tokens/permissions
CEF allows Chromium to call HardenTokenIntegrityLevelPolicy, which
basically screws with our security descriptors, preventing our new
window capture method from working properly.  This is because even
though we already told CEF that we don't want sandboxing, it's basically
initializing it anyway for whatever reason.  So, instead of allowing it
to do that, just disable sandboxing directly.

Note that this is a massive hack.  But it's a really nice hack.

(I take full responsibility for this, this is just so much more clean
and simple than the alternatives.  -Jim)

Closes obsproject/obs-browser#204