Rename obs-browser-page directory to obs-browser-page-main. Close #239
As discussed in #239, `cmake` is being instructed to build the `obs-browser` binary using the same name as a directory that already exists.

This pull request renames the `obs-browser-page` directory to `obs-browser-page-main` and updates `CMakeLists.txt` accordingly.

This fixes the FTBFS report in #239 using Ubuntu 20.04, `gcc` 9.3.0 and `cmake` 3.16.

While I did not encounter this issue using Ubuntu 18.04, `gcc` 8.4.0 and ``cmake` 3.10; I've tested building on that platform with this fix applied and compilation continues to complete successfully.