Fix Shared Texture option in cmake-gui
When using set without specifying the CACHE argument, a normal/local
variable is set to the specified value. This results in the GUI showing
the CACHE variable value (OFF), but the rest of the CMake script checks
the normal variable value first. Thus, the script works as if
SHARED_TEXTURE_SUPPORT_ENABLED is set to ON, even though cmake-gui shows
it set to OFF.