Fix CEF locale path on *nix platforms
This fixes an issue where the browser panel context menu would
have blank text, making it near-impossible to use.
This issue does not occur on Windows. macOS uses a special path
for Framework locales, and handles *most* languages without issue.

This solution does not work if the locale path is relative.