Expose OBS controls to browser sources
Exposes functions of the Frontend API to the webpage of browser sources
via JS, like the ability to start and stop recordings etc.
Adds a menu to the source properties that lets the user decide which
permissions the webpage is given, grouped into different levels (none,
read-only, basic, advanced, all).