Pass on JS function arguments to BrowserClient
When functions were called from JS, the only argument that would
previously be respected is the callback function, if it exists,
additional arguments or non-function arguments would be ignored.
This makes the BrowserApp::Execute function pass all other arguments on
to the BrowserClient::OnProcessMessageReceived function, so that they
can be used.

This part of the message is only important for people adding new
functions, so they don't have to dig too deep into the code:

Due to how callback functions are handled, this needs to be used with
caution: The callbackId is (like it was before) ALWAYS the first
argument, no matter whether a function was actually passed. If it
wasn't, that Int will be zero, but still defined.
All non-callback-function arguments start at index 1 of the
message->GetArgumentList. This means that, if no callback was passed by
JS, the function args will have their index increased by 1 compared to
their index in the caller function. If a callback was given in JS as the
first argument, the index will be the same as in JS (since
non-callback-arguments would start at the second position there as well)