Ignore overlay texture when hardware accelerated
This fixes an issue where opening an <input type="date"> would result in
flickering between the base texture & a smaller texture of the date
picker. This brings behaviour in line with non-accelerated rendering.

Ideally PET_POPUP textures (both accelerated & not) should be properly
stored and rendered on top of the bs->texture displayed to the user.

For more info see #148