Fix hard lock when destroying browser sources
It is possible that while in the graphics mutex, a browser source will
be destroyed, and will call the synchronous DestroyBrowser, which then
will wait on the CEF thread. But the CEF thread can then get stuck
waiting on the graphics thread because a render task may be queued
within the CEF thread before that destroy is called.

To prevent that, before calling DestroyBrowser synchronously, mark the
browser source for destruction, and don't allow those conflicting tasks
to be executed.

Fixes obsproject/obs-browser#333