Fix Windows crash if shared texture is unavailable
Users on dual-GPU systems, broken drivers, or virtual machines running
Windows would experience a crash when trying to load a hardware
accelerated browser source. This brings the behaviour of 4638 in line
with 3770, which is silently fail & log an error.

Ideally, OBS should destroy the browser and recreate it with hardware
acceleration disabled.