cmake: Fix handling of debug wrapper libraries
Specifying separate targets for debug configuration (as well as using
"debug" and "optimized" keywords) is legacy behavior.

The correct way is to specify `_DEBUG` variants of the imported target
properties, which CMake will pick up automatically if the project is
built with debug configuration.

This commit forward-fixes the update on obs-studio, by still using
the legacy behavior if the legacy target is present.