Fix handling frontend JS events
When handling frontend events, DispatchJSEvent expects JSON passed as a
std::string. An empty string will cause the event to never be
dispatched. An empty nlohmann::json object converted to a std::string is
represented by the string "null". Passing this exact string as the JSON
string to DispatchJSEvent makes the events work. This is probably
because an empty unquoted string by itself is not valid JSON. Two
double-quotes with nothing inside them or the unquoted string "null" are
considered valid JSON.