Add support for setting zoom for browser sources
Previously, the only way to scale up a page was to use custom CSS,
which may or may not work well depending on how the CSS rules of the
page being displayed were written, or to transform the browser source,
which would be blurry.

To make it possible to do this properly, I've implemented support for
setting the zoom level of the browser, equivalent to pressing Ctrl-+ /
Ctrl-- in a browser.

The range of [-6, 9] comes from kPresetZoomFactorsArray in Chromium.
Zoom levels are indexes into this array, offset so that a 0 zoom level
is 1.0.