Don't loop Cef exit while shutting down if task post fails
If the `manager_thread` exits early, there will be no message loop to
post the task to, and a permanent deadlock would occur. Rather than
guaranteeing a deadlock, this at least lowers the odds of one.

A future implementation might be more aware of the run state of the
manager thread and handle these exit scenarios more explicitly.