Log error if CefInitialize fails
So it turns out CefInitialize is a boolean return... Anyway, let's log
failures so that we can at least begin to track them down. This change
also returns early, but that does not make any of what we're doing any
safer. Returning early currently only has the effect of safely
disabling browser panels.

More work should be done to fully lock out the plugin if CefInitialize
fails. Such behavior is required according to the CEF documentation.

See the CEF cef_resultcode_t enum for the definition of the exit codes.